app/pyproject.toml
[tool.poetry]
name = "payment"
version = "0.1.0"
description = ""
authors = ["Charles Tenorio <charles.tenorio.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
asaas-sdk = "^3.2.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
passlib = "^1.7.4"
pytz = "^2024.1"
sqlalchemy = "^2.0.26"
pydantic = {extras = ["email"], version = "^2.6.1"}
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
